requirements.txt
line-bot-sdk==2.4.1
numpy==1.22.1
python-dotenv==0.21.1
Flask==2.2.2
hyperdb-python==0.1.2
pandas==2.0.0
openai==0.27.4